test(analyze): stub-provider test for BedrockAuth credential-error mapping
Extract the credential-error heuristic into `classify_bedrock_error` and
`explain_with_bedrock_provider` so the mapping can be exercised offline via a
`CredentialErrorProvider` stub. Add three tests: an async end-to-end test that
injects the stub through `explain_with_bedrock_provider` and asserts
`BedrockAuth` is returned; a sync unit test covering the classify heuristic for
multiple credential-keyword forms; and a sync negative test confirming generic
errors map to `Chat`, not `BedrockAuth`. Split tests to `explain_tests.rs` to
bring `explain.rs` under the 500-line cap and ratchet the allowlist entry down.

Closes #951
